+++ test/mix/tasks/bonny.gen.controller_test.exs
@@ -15,6 +15,7 @@ defmodule Mix.Tasks.Bonny.Gen.ControllerTest do
     [
       crd_no_args: ~w(McController y Memcached) |> input_args(),
       crd_arg_controller: ~w(y Memcached) |> input_args(),
+      crd_arg_controller_inv: ~w(y Mem_Cached Memcached) |> input_args(),
       crd_arg_controller_version: ~w(y Memcached) |> input_args(),
       deployment_no_args: ["DeplController", "n", @choice_resource_deployment] |> input_args(),
       deployment_arg_controller: ["n", @choice_resource_deployment] |> input_args(),
@@ -51,6 +52,26 @@ defmodule Mix.Tasks.Bonny.Gen.ControllerTest do
       assert output =~ "defmodule Bonny.Controller.McController do"
       assert output =~ "defmodule Bonny.Controller.McControllerTest do"
       assert output =~ "defmodule Bonny.Test.API.V1.Memcached do"
+    end
+
+    test "Asks for controller name again if it is invalid.",
+         %{crd_no_args: input} do
+      output =
+        capture_io([input: "Invalid_Controller_Name\n" <> input], fn ->
+          Controller.run(["--out", "-"])
+        end)
+
+      assert output =~ "not a valid Elixir module name!"
+    end
+
+    test "Asks for CRD name again if it is invalid.",
+         %{crd_arg_controller_inv: input} do
+      output =
+        capture_io([input: input], fn ->
+          Controller.run(["--out", "-", "McController"])
+        end)
+
+      assert output =~ "is not a valid kubernetes kind!"
     end
 
     test "generates a CRD controller and a Version module module when controller name is passed.",
@@ -141,7 +162,11 @@ defmodule Mix.Tasks.Bonny.Gen.ControllerTest do
       _modules =
         output
         |> String.replace(~r/^.*?defmodule/s, "defmodule")
-        |> Code.compile_string()
+        |> String.replace(
+          ~r/.*Don't forget to add the controller to the list of controllers in your application config.*/,
+          ""
+        )
+        |> Code.eval_string()
     end
 
     test "generated modules compile (deployment)", %{deployment_no_args: input} do
@@ -153,7 +178,11 @@ defmodule Mix.Tasks.Bonny.Gen.ControllerTest do
       _modules =
         output
         |> String.replace(~r/^.*?defmodule/s, "defmodule")
-        |> Code.compile_string()
+        |> String.replace(
+          ~r/.*Don't forget to add the controller to the list of controllers in your application config.*/,
+          ""
+        )
+        |> Code.eval_string()
     end
 
     test "generated modules compile (sealed secret)", %{sealed_secret_no_args: input} do
@@ -165,6 +194,10 @@ defmodule Mix.Tasks.Bonny.Gen.ControllerTest do
       _modules =
         output
         |> String.replace(~r/^.*?defmodule/s, "defmodule")
+        |> String.replace(
+          ~r/.*Don't forget to add the controller to the list of controllers in your application config.*/,
+          ""
+        )
         |> Code.compile_string()
     end
 
@@ -192,6 +225,15 @@ defmodule Mix.Tasks.Bonny.Gen.ControllerTest do
         end)
 
       assert output =~ "defmodule Bonny.Controller.McControllerTest do"
+    end
+
+    test "prints the help message" do
+      output =
+        capture_io(fn ->
+          catch_exit(Controller.run(["-h"]))
+        end)
+
+      assert output =~ "mix bonny.gen.controller"
     end
   end
 end