+++ test/support/controllers/test_resource_v3_controller.ex
@@ -7,7 +7,6 @@ defmodule TestResourceV3Controller do
   require CRD
 
   use Bonny.ControllerV2,
-    for_resource: CRD.build_for_controller!(names: CRD.kind_to_names("TestResourceV3")),
     skip_observed_generations: true
 
   @impl Bonny.ControllerV2